update queries to include user_ID

+++ assign_seat.php
@@ -1,4 +1,5 @@
 <?php
+require_once 'includes/check_session.inc.php';
 // assign_seat.php
 
 $servername = "localhost";
@@ -15,6 +16,7 @@ if ($conn->connect_error) {
 }
 
 // Get data from the AJAX request
+$user_ID = $_SESSION['user_id'];
 $classId = isset($_GET['classId']) ? $_GET['classId'] : null;
 $seatNumber = isset($_GET['seatNumber']) ? $_GET['seatNumber'] : null;
 $studentID = isset($_GET['studentID']) ? $_GET['studentID'] : null;
@@ -25,9 +27,9 @@ $response = array();
 // Perform UPSERT operation (replace INTO is MySQL's UPSERT)
 $sql = "
 START TRANSACTION;
-DELETE FROM tb_seatplan WHERE student_ID = $studentID AND class_ID = $classId;
-INSERT INTO tb_seatplan (class_ID, seat_number, student_ID)
-VALUES ($classId, '$seatNumber', $studentID);
+DELETE FROM tb_seatplan WHERE user_ID = $user_ID AND student_ID = $studentID AND class_ID = $classId;
+INSERT INTO tb_seatplan (user_ID, class_ID, seat_number, student_ID)
+VALUES ($user_ID, $classId, '$seatNumber', $studentID);
 COMMIT;
 ";
